refactor: replace axios with fetch. put handleCommand to a new file
