+++ packages/container/package.json
@@ -8,7 +8,6 @@
     "watch": ""
   },
   "dependencies": {
-    "axios": "^1.8.2",
     "minimist": "^1.2.8"
   }
 }
